Improve beta request states handling for organizations
