+++ app/models/owner.js
@@ -2,20 +2,19 @@ import Model from 'ember-data/model';
 import attr from 'ember-data/attr';
 import { belongsTo } from 'ember-data/relationships';
 import { computed } from '@ember/object';
-import { reads, or } from '@ember/object/computed';
+import { reads, or, notEmpty } from '@ember/object/computed';
 import { inject as service } from '@ember/service';
 import config from 'travis/config/environment';
 import dynamicQuery from 'travis/utils/dynamic-query';
-import withBetaMigrationRequests from 'travis/mixins/model/owner/with-beta-migration-requests';
 
 const { profileReposPerPage: limit } = config.pagination;
 
-export default Model.extend(withBetaMigrationRequests, {
+export default Model.extend({
   features: service(),
   accounts: service(),
   raven: service(),
-  ajax: service(),
   store: service(),
+  tasks: service(),
 
   name: attr('string'),
   login: attr('string'),
@@ -62,6 +61,24 @@ export default Model.extend(withBetaMigrationRequests, {
       limit, offset, custom: { owner, type, },
     }, { live: false });
   },
+
+  fetchBetaMigrationRequests() {
+    return this.tasks.fetchBetaMigrationRequestsTask.perform();
+  },
+
+  migrationBetaRequests: computed('tasks.fetchBetaMigrationRequestsTask.lastSuccessful.value.[]', 'id', function () {
+    const requests = this.tasks.fetchBetaMigrationRequestsTask.get('lastSuccessful.value') || [];
+    return requests.filter(request =>
+      this.isUser && request.ownerId == this.id || request.organizations.mapBy('id').includes(this.id)
+    );
+  }),
+
+  isMigrationBetaRequested: notEmpty('migrationBetaRequests'),
+
+  isMigrationBetaAccepted: computed('migrationBetaRequests.@each.acceptedAt', function () {
+    const migrationBetaRequests = this.migrationBetaRequests || [];
+    return migrationBetaRequests.isAny('acceptedAt');
+  }),
 
   subscriptionError: reads('accounts.subscriptionError'),
